requirements.txt
fastapi
httpx
pydantic
python-dotenv
supabase
edge-tts
google-generativeai
uvicorn
